development

+++ src/main/java/data/SplitDataImpl.java
@@ -74,9 +74,16 @@ public class SplitDataImpl implements SplitDataService {
 		String sql5 = "SELECT * FROM merge WHERE final_id = " + id;
 		rSet = statement.executeQuery(sql5);
 		if(!rSet.next()) {
-			String sql6 = "DELETE FROM report WHERE id = " + id;
-			int i = statement.executeUpdate(sql6);
-			if(i==0) return false;
+			int origin = getOrigin(id);
+			if(origin == 0) {
+				String sql7 = "UPDATE report SET merge = 0 WHERE id = " + id;
+				statement.executeUpdate(sql7);
+			}
+			else {
+				String sql6 = "DELETE FROM report WHERE id = " + id;
+				int i = statement.executeUpdate(sql6);
+				if(i==0) return false;				
+			}
 		}
 		DBManager.stopAll(rSet, statement, connection);
 		return true;
